splitting Blackoil related to FlowProblemBlackoil
so FlowProblem can be used for compositional or other setting
