+++ CMakeLists_files.cmake
@@ -756,6 +756,7 @@ list (APPEND PUBLIC_HEADER_FILES
   opm/simulators/flow/FlowGenericVanguard.hpp
   opm/simulators/flow/FlowMain.hpp
   opm/simulators/flow/FlowProblem.hpp
+  opm/simulators/flow/FlowProblemBlackoil.hpp
   opm/simulators/flow/FlowProblemParameters.hpp
   opm/simulators/flow/FlowProblemProperties.hpp
   opm/simulators/flow/FlowUtils.hpp
