+++ opm/simulators/flow/BaseAquiferModel.hpp
@@ -108,6 +108,13 @@ public:
                      unsigned) const
     { }
 
+
+    void addToSource(RateVector&,
+                     unsigned,
+                     unsigned) const
+    { }
+
+
     /*!
      * \brief This method is called after each Newton-Raphson successful iteration.
      *
